+++ src/main/kotlin/domain/square/mine/Mine.kt
@@ -1,4 +1,4 @@
-package domain.mine
+package domain.square.mine
 
 import domain.position.Position
