+++ src/main/kotlin/domain/square/mine/MineFactory.kt
@@ -1,4 +1,4 @@
-package domain.mine
+package domain.square.mine
 
 import domain.position.PositionIdFactory
 import domain.position.RandomPositionIdFactory
