+++ src/test/kotlin/domain/MinesTest.kt
@@ -1,7 +1,7 @@
 package domain
 
-import domain.mine.Mine
-import domain.mine.Mines
+import domain.square.mine.Mine
+import domain.square.mine.Mines
 import domain.position.Position
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
